+++ src/composables/index.ts
@@ -1,11 +1,11 @@
-export * from '@inkline/inkline/composables/click-outside';
-export * from '@inkline/inkline/composables/collapsible';
-export * from '@inkline/inkline/composables/inkline';
-export * from '@inkline/inkline/composables/isServer';
-export * from '@inkline/inkline/composables/linkable';
-export * from '@inkline/inkline/composables/componentColor';
-export * from '@inkline/inkline/composables/componentSize';
-export * from '@inkline/inkline/composables/form';
-export * from '@inkline/inkline/composables/popup';
-export * from '@inkline/inkline/composables/toast';
-export * from '@inkline/inkline/composables/validation';
+export * from '@inkline/inkline/composables/useClickOutside';
+export * from '@inkline/inkline/composables/useCollapsible';
+export * from '@inkline/inkline/composables/useComponentColor';
+export * from '@inkline/inkline/composables/useComponentSize';
+export * from '@inkline/inkline/composables/useForm';
+export * from '@inkline/inkline/composables/useInkline';
+export * from '@inkline/inkline/composables/useIsServer';
+export * from '@inkline/inkline/composables/useLinkable';
+export * from '@inkline/inkline/composables/usePopupControl';
+export * from '@inkline/inkline/composables/useToast';
+export * from '@inkline/inkline/composables/useValidation';
