+++ src/composables/useClickOutside.ts
