+++ src/composables/useCollapsible.ts
